fix for share function.
This performs the following changes:
- apply a variant of shareReplay.
- map the memoized observable to the source observable

The first change ensures that the selector runs only when refCount go from 0 to 1
and that subscribers will always receive the latest value from the source.

The second change ensures that if a memoized selector is called with different source observables, it returns distinct results.